+++ lib/fog/azurerm/requests/compute/create_availability_set.rb
@@ -9,11 +9,11 @@ module Fog
           resource_group = availability_set_params[:resource_group]
           fault_domain_count = availability_set_params[:platform_fault_domain_count]
           update_domain_count = availability_set_params[:platform_update_domain_count]
-          is_managed = availability_set_params[:is_managed].nil? ? false : availability_set_params[:is_managed]
+          use_managed_disk = availability_set_params[:use_managed_disk].nil? ? false : availability_set_params[:use_managed_disk]
 
           msg = "Creating Availability Set '#{name}' in #{location} region."
           Fog::Logger.debug msg
-          avail_set_params = get_availability_set_properties(location, fault_domain_count, update_domain_count, is_managed)
+          avail_set_params = get_availability_set_properties(location, fault_domain_count, update_domain_count, use_managed_disk)
 
           begin
             availability_set = @compute_mgmt_client.availability_sets.create_or_update(resource_group, name, avail_set_params)
@@ -25,18 +25,18 @@ module Fog
         end
 
         # create the properties object for creating availability sets
-        def get_availability_set_properties(location, fault_domain_count, update_domain_count, is_managed)
+        def get_availability_set_properties(location, fault_domain_count, update_domain_count, use_managed_disk)
           avail_set = Azure::ARM::Compute::Models::AvailabilitySet.new
           avail_set.location = location
-          avail_set.sku = create_availability_set_sku(is_managed)
+          avail_set.sku = create_availability_set_sku(use_managed_disk)
           avail_set.platform_fault_domain_count = fault_domain_count.nil? ? FAULT_DOMAIN_COUNT : fault_domain_count
           avail_set.platform_update_domain_count = update_domain_count.nil? ? UPDATE_DOMAIN_COUNT : update_domain_count
           avail_set
         end
 
-        def create_availability_set_sku(is_managed)
+        def create_availability_set_sku(use_managed_disk)
           sku = Azure::ARM::Compute::Models::Sku.new
-          sku.name = is_managed ? AS_SKU_ALIGNED : AS_SKU_CLASSIC
+          sku.name = use_managed_disk ? AS_SKU_ALIGNED : AS_SKU_CLASSIC
           sku
         end
       end